Separated UI and JS FrameCallbackRegistries
