+++ src/reanimated2/frameCallback/FrameCallbackRegistryUI.ts
@@ -1,6 +1,6 @@
-import { runOnUI } from '../';
+import { runOnUI } from '..';
 
-export interface FrameCallbackRegistryUI {
+export default interface FrameCallbackRegistryUI {
   frameCallbackRegistry: Map<number, () => void>;
   frameCallbackActive: Set<number>;
   isFrameCallbackRunning: boolean;
@@ -10,7 +10,7 @@ export interface FrameCallbackRegistryUI {
   manageStateFrameCallback: (frameCallbackId: number, state: boolean) => void;
 }
 
-const prepareUIRegistry = runOnUI(() => {
+export const prepareUIRegistry = runOnUI(() => {
   'worklet';
 
   const frameCallbackRegistry: FrameCallbackRegistryUI = {
@@ -62,46 +62,3 @@ const prepareUIRegistry = runOnUI(() => {
 
   global._frameCallbackRegistry = frameCallbackRegistry;
 });
-
-export default class FrameCallbackRegistryJS {
-  private nextCallbackId = 0;
-
-  constructor() {
-    if (global._frameCallbackRegistry === undefined) {
-      prepareUIRegistry();
-    }
-  }
-
-  registerFrameCallback(callback: () => void): number {
-    if (!callback) {
-      return -1;
-    }
-
-    const callbackId = this.nextCallbackId;
-    this.nextCallbackId++;
-
-    runOnUI(() => {
-      'worklet';
-      global._frameCallbackRegistry.registerFrameCallback(callback, callbackId);
-    })();
-
-    return callbackId;
-  }
-
-  unregisterFrameCallback(frameCallbackId: number): void {
-    runOnUI(() => {
-      'worklet';
-      global._frameCallbackRegistry.unregisterFrameCallback(frameCallbackId);
-    })();
-  }
-
-  manageStateFrameCallback(frameCallbackId: number, state: boolean): void {
-    runOnUI(() => {
-      'worklet';
-      global._frameCallbackRegistry.manageStateFrameCallback(
-        frameCallbackId,
-        state
-      );
-    })();
-  }
-}
